# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CommsPro - Professional Communication Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized

  # ─── DASHBOARD ──────────────────────────────────────────────────────────────
...
  /contacts:
    get:
      operationId: getContacts
      tags: [contacts]
      summary: List contacts
      responses:
        "200":
          description: Contacts list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contact"
    post:
      operationId: createContact
      tags: [contacts]
      summary: Create a contact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
  /contacts/{id}:
    get:
      operationId: getContact
      tags: [contacts]
      summary: Get a contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
    patch:
      operationId: updateContact
      tags: [contacts]
      summary: Update a contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
    delete:
      operationId: deleteContact
      tags: [contacts]
      summary: Delete a contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ─── CALLS ──────────────────────────────────────────────────────────────────
...
  /calls/outbound:
    post:
      operationId: makeCall
      tags: [calls]
      summary: Make an outbound call
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CallInput"
      responses:
        "201":
          description: Call initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Call"
  /calls/{id}:
    get:
      operationId: getCall
      tags: [calls]
      summary: Get call details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Call details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Call"
  /calls/{id}/end:
    post:
      operationId: endCall
      tags: [calls]
      summary: End a call
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Call ended
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Call"
  /calls/{id}/answer:
    post:
      operationId: answerCall
      tags: [calls]
      summary: Answer an incoming call
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Call answered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Call"
  /calls/{id}/reject:
    post:
      operationId: rejectCall
      tags: [calls]
      summary: Reject an incoming call
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Call rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Call"

  # ─── SMS ────────────────────────────────────────────────────────────────────
...
  /agents:
    get:
      operationId: getAgents
      tags: [agents]
      summary: List AI agents
      responses:
        "200":
          description: Agents list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Agent"
    post:
      operationId: createAgent
      tags: [agents]
      summary: Create an AI agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
  /agents/{id}:
    get:
      operationId: getAgent
      tags: [agents]
      summary: Get an AI agent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Agent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
    patch:
      operationId: updateAgent
      tags: [agents]
      summary: Update an AI agent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
    delete:
      operationId: deleteAgent
      tags: [agents]
      summary: Delete an AI agent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /agents/{id}/toggle:
    post:
      operationId: toggleAgent
      tags: [agents]
      summary: Enable or disable an AI agent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentToggle"
      responses:
        "200":
          description: Toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
  /agents/{id}/tasks:
    get:
      operationId: getAgentTasks
      tags: [agents]
      summary: Get tasks assigned to an agent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AgentTask"
    post:
      operationId: assignAgentTask
      tags: [agents]
      summary: Assign a task to an agent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentTaskInput"
      responses:
        "201":
          description: Task assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentTask"
  /agents/{id}/activity:
    get:
      operationId: getAgentActivity
      tags: [agents]
      summary: Get activity log for an agent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Activity log
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AgentActivityItem"

  # ─── COMMUNICATION LOGS ─────────────────────────────────────────────────────
...
components:
  schemas:
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    # ── Auth ──────────────────────────────────────────────────────────────────
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 1
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string

    # ── User ──────────────────────────────────────────────────────────────────
    User:
      type: object
      required: [id, name, email, role, isOnline, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, agent, viewer]
        isOnline:
          type: boolean
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...
    Contact:
      type: object
      required: [id, name, phone, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        company:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    ContactInput:
      type: object
      required: [name, phone]
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        company:
          type: string
        notes:
          type: string
    ContactUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        company:
          type: string
        notes:
          type: string

    # ── Call ──────────────────────────────────────────────────────────────────
    Call:
      type: object
      required: [id, direction, from, to, status, createdAt]
      properties:
        id:
          type: integer
        twilioCallSid:
          type: ["string", "null"]
        direction:
          type: string
          enum: [inbound, outbound]
        from:
          type: string
        to:
          type: string
        status:
          type: string
          enum: [queued, ringing, in-progress, completed, busy, failed, no-answer, canceled]
        duration:
          type: ["integer", "null"]
        recordingUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CallInput:
      type: object
      required: [to]
      properties:
        to:
          type: string
        from:
          type: string

    # ── SMS ───────────────────────────────────────────────────────────────────
...
    Agent:
      type: object
      required: [id, name, type, status, isEnabled, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [call-handler, sms-responder, whatsapp-bot, scheduler, monitor]
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [idle, active, busy, error]
        isEnabled:
          type: boolean
        taskCount:
          type: integer
        lastActiveAt:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    AgentInput:
      type: object
      required: [name, type]
      properties:
        name:
          type: string
        type:
          type: string
          enum: [call-handler, sms-responder, whatsapp-bot, scheduler, monitor]
        description:
          type: string
    AgentUpdate:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [call-handler, sms-responder, whatsapp-bot, scheduler, monitor]
        description:
          type: string
    AgentToggle:
      type: object
      required: [isEnabled]
      properties:
        isEnabled:
          type: boolean
    AgentTask:
      type: object
      required: [id, agentId, title, status, createdAt]
      properties:
        id:
          type: integer
        agentId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, running, completed, failed]
        priority:
          type: string
          enum: [low, medium, high]
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: ["string", "null"]
    AgentTaskInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
        description:
          type: string
        priority:
          type: string
          enum: [low, medium, high]
    AgentActivityItem:
      type: object
      required: [id, agentId, event, createdAt]
      properties:
        id:
          type: integer
        agentId:
          type: integer
        event:
          type: string
        detail:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time

    # ── Dashboard ─────────────────────────────────────────────────────────────
...