# lib/api-zod/src/generated/types/agentUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * CommsPro - Professional Communication Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { AgentUpdateType } from './agentUpdateType';

export interface AgentUpdate {
  name?: string;
  type?: AgentUpdateType;
  description?: string;
}